moved pdf into own file, fixed some pdf bugs
