+++ blamo/lib/PDF/pdf_dummy_data.dart
@@ -1,0 +1,39 @@
+// SAMPLE LOG DATA
+String objectID = "1";
+String testType = "SPT";
+String project = "I-5 @ OR214 Interchange (Woodburn) Development Sec.";
+String number = "PE000559";
+String client = "Oregon Department of Transportation";
+String lat = "550129.7734";
+String long = "7590750.9362";
+String location = "MP 33.75";
+String elevationDatum = "[Elevation Datum]";
+String boreholeID = "12518-01";
+String startDate = "5/1/2010";
+String endDate = "5/1/2010";
+String surfaceElevation = "186.51";
+String contractor = "Adonis - Western States";
+String method = "MUD ROTARY - AUTO HAMMER";
+String loggedBy = "Castelli";
+String checkedBy = "HSK / GAF / JFF";
+
+//TEST SAMPLE DATA
+String beginTestDepth = "0";
+String endTestDepth = "-2.5";
+String soilType = "ML";
+String description = "SILT with trace to some Sand";
+String moistureContent = "38-31";
+String dryDensity = "dry density";
+String liquidLimit = "liquid limit";
+String plasticLimit = "plastic limit";
+String fines = "fines";
+String blows1 = "5";
+String blows2 = "7";
+String blows3 = "11"; 
+String blowCount = "18";
+
+//UNIT SAMPLE DATA
+String beginUnitDepth = "0";
+String endUnitDepth = "-2.5";
+String unitDescription = "Sandy GRAVEL (Shoulder Aggregate), GP";
+String unitMethods = "Unit Drilling Method";
